requirements.txt
# Core dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
Pillow>=8.0.0
tqdm>=4.62.0

# Vision-Language Models
clip @ git+https://github.com/openai/CLIP.git
open-clip-torch>=2.20.0

# Data processing
pandas>=1.3.0
csv
pathlib

# Dimensionality reduction
umap-learn>=0.5.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
tabulate>=0.8.9

# Optimization
optuna>=3.0.0

# Optional but recommended
jupyter>=1.0.0
ipywidgets>=7.6.0